+++ bootstrap/bootstrap.sh
@@ -29,6 +29,7 @@ if [[ "$K3SFRAMEWORK_TYPE" == "server" ]]
 then
   curl -L http://dl.k8s.io/release/$KUBERNETES_VERSION/bin/linux/amd64/kubectl > $MESOS_SANDBOX/kubectl
   curl https://raw.githubusercontent.com/AVENTER-UG/mesos-m3s/dev/bootstrap/server > $MESOS_SANDBOX/server
+  curl https://raw.githubusercontent.com/AVENTER-UG/mesos-m3s/dev/bootstrap/update.sh > $MESOS_SANDBOX/update
   chmod +x $MESOS_SANDBOX/kubectl
   chmod +x $MESOS_SANDBOX/server
   exec $MESOS_SANDBOX/server &